Redesign some aspects of the SettingsManagers, JsonFormats & improve legacy config upgrading
- Removed methods that accessed commands by their array index (This was fucking stupid and shouldn't have existed in the first place. Some solice in that it went unused.)
- Changed the role management method names to something more sensible (I could never remember which way round they were before!)
- Changed the permission check method to "IsPermitted"
- Added permission check via command key

- SettingsJsons now inherit from a super class
- CommandConfigJsons now inherit from a superclass

- Registered commands in command register no longer need to be added so that the newest command is at the bottom. They can now be added willy nilly in there.

- Various small optimisations and improvements

All references have been updated to reflect these changes.

Most of my time here was dedicated to finding some way to connect the GuildSettingsManager and GlobalSettingsManager, as many of the methods are similar, only differing in the type of CommandConfigJson/SettingsJson they use, however with these differences and the mixture of static/instanced and guildIDs, I couldn't find a workaround as yet.