Initial implementation of new Command Config Menu for Guilds.
Much like the game settings, it's rather messy and will likely break in the future as more commands get added due to embed size limits.
I considered just leaving a prompt to use /help [category all, but it just seemed like poor UX, so I'll look for an alternative to showing the data.