Complete code tidy up and UX improvements to Command Config
Bloody hell, why on Earth I implemented this all into a single function I'll never understand, and not sure I want to. But that nightmare is behind us now.

- De fucked the code
- Removed the command list (I'm assuming the user knows the command they want to edit if they're going for single command modification)
	- It would have only caused probelems as it grew larger, as it would eventually extend Discord's embed size limit.
- Fixed the issue where roles could be granted permissions multiple times
- Fixed a small graphical issue where roles had no spacing between them
- Removed the arbitrary nesting of loops in role application
- Added the ability to add roles by ID (Useful for when two roles have the same name, and the user only wants to apply one)
	- An alternate fix would be to acknowledge case differences, but quite frankly, convinience to users outweighs the niche scenario when this would be useful
- Added exception throwing to be consistent with other config pages
- Some other stuff(?)

Future plans are to document this and add a system to modify by role, whereby the user can select one role, and then define multiple commands to grant access to. (Useful for when new roles are added after the bot has already been set up)