Add load class loading
I really need to come up with a better name for this, huh? Anyway, this class is one that denotes the methods within should be executed when the module is loaded, as opposed to when a command request is issued.

I've also started work on more sophisticated handling of errors and issues.